requirements.txt
streamlit==1.28.0
numpy==1.24.3
opencv-python-headless==4.8.1.78
Pillow==10.1.0
pandas==2.0.3
matplotlib==3.7.3
torch==2.1.0
torchvision==0.16.0
ultralytics==8.0.196
scikit-learn==1.3.1
scipy==1.11.3
